requirements.txt
torch
argparse
numpy
transformers
nltk
opencv_python
pandas
Pillow
scikit_learn